cleanup and fix
Clean up the code to make it more readable, for example, use only println on the HTTP request.

Add in ability to keep the data if the connection to pompeii fails.  Send to nginx instead of php and fix the issue with 499 in the nginx logs because the connection closes too fast.